backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventory-management-system"
version = "0.1.0"
description = "智能库存管理系统 - 后端服务"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Inventory Team" }
]

dependencies = [
    # Web 框架
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "python-multipart>=0.0.9,<1.0.0",

    # 数据库
    "sqlalchemy[asyncio]>=2.0.30,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    "greenlet>=3.0.0,<4.0.0",

    # 缓存与任务队列
    "redis>=5.0.0,<6.0.0",
    "celery[redis]>=5.4.0,<6.0.0",

    # 数据处理
    "pandas>=2.2.0,<3.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "numpy>=1.26.0,<3.0.0",

    # 统计与科学计算
    "statsmodels>=0.14.0,<1.0.0",
    "scipy>=1.13.0,<2.0.0",

    # AI / LLM
    "langchain>=0.2.0,<1.0.0",
    "langchain-openai>=0.1.0,<1.0.0",
    "openai>=1.30.0,<2.0.0",

    # HTTP 客户端
    "httpx>=0.27.0,<1.0.0",

    # 重试机制
    "tenacity>=8.3.0,<9.0.0",

    # 邮件发送
    "aiosmtplib>=3.0.0,<4.0.0",

    # WebSocket
    "websockets>=12.0,<14.0",

    # 配置管理
    "pydantic-settings>=2.3.0,<3.0.0",
    "pydantic>=2.7.0,<3.0.0",

    # 认证与安全
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",

    # 数据验证
    "email-validator>=2.1.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "httpx>=0.27.0,<1.0.0",
    "ruff>=0.5.0,<1.0.0",
    "mypy>=1.10.0,<2.0.0",
    "pre-commit>=3.7.0,<4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["app"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
